TestScriptAnalyzer/requirements.txt
pytest>=7.0
playwright>=1.48.0
pytest-playwright>=0.5.0
pyyaml>=6.0
tabulate>=0.9
openpyxl>=3.0.0
pandas>=1.5.0
